minor, move JobInfoConverter to kylin-engine-mr
